+++ stitch/bindings_test.js
@@ -468,7 +468,7 @@ describe('Bindings', function() {
             }]);
         });
         it('connect to invalid port range', function() {
-            expect(() => foo.connect(true, bar)).to
+            expect(() => foo.allowFrom(bar, true)).to
                 .throw('Input argument must be a number or a Range');
         });
         it('allow connections to publicInternet', function() {
@@ -508,80 +508,6 @@ describe('Bindings', function() {
                     `and not from a Container or other object.`);
         });
     });
-    describe('Connect', function() {
-        // This test runs all of the same tests as AllowFrom, but uses the
-        // deprecated connect() function rather than the newer allowFrom()
-        // function. We can remove this as soon as we remove support for
-        // connect().
-        let foo;
-        let bar;
-        beforeEach(function() {
-            foo = new Service('foo', []);
-            bar = new Service('bar', []);
-            deployment.deploy([foo, bar]);
-        });
-        const checkConnections = function(expected) {
-            const {connections} = deployment.toQuiltRepresentation();
-            expect(connections).to.have.lengthOf(expected.length)
-                .and.containSubset(expected);
-        };
-        it('port', function() {
-            foo.connect(new Port(80), bar);
-            checkConnections([{
-                from: 'foo',
-                to: 'bar',
-                minPort: 80,
-                maxPort: 80,
-            }]);
-        });
-        it('port range', function() {
-            foo.connect(new PortRange(80, 85), bar);
-            checkConnections([{
-                from: 'foo',
-                to: 'bar',
-                minPort: 80,
-                maxPort: 85,
-            }]);
-        });
-        it('connect to invalid port range', function() {
-            expect(() => foo.connect(true, bar)).to
-                .throw('Input argument must be a number or a Range');
-        });
-        it('connect to publicInternet', function() {
-            foo.connect(80, publicInternet);
-            checkConnections([{
-                from: 'foo',
-                to: 'public',
-                minPort: 80,
-                maxPort: 80,
-            }]);
-        });
-        it('connect from publicInternet', function() {
-            publicInternet.connect(80, foo);
-            checkConnections([{
-                from: 'public',
-                to: 'foo',
-                minPort: 80,
-                maxPort: 80,
-            }]);
-        });
-        it('connect to publicInternet port range', function() {
-            expect(() => foo.connect(new PortRange(80, 81), publicInternet)).to
-                .throw('public internet can only connect to single ports ' +
-                    'and not to port ranges');
-        });
-        it('connect from publicInternet port range', function() {
-            expect(() => publicInternet.connect(new PortRange(80, 81), foo)).to
-                .throw('public internet can only connect to single ports ' +
-                    'and not to port ranges');
-        });
-        it('connect to non-service', function() {
-            expect(() => foo.connect(10, 10)).to
-                .throw(`Services can only connect to other services. ` +
-                    `Check that you're connecting to a service, and not ` +
-                    `to a Container or other object.`);
-        });
-    });
     describe('Vet', function() {
         let foo;
         const deploy = () => deployment.toQuiltRepresentation();
@@ -599,7 +525,7 @@ describe('Bindings', function() {
             foo.placeOn({
                 floatingIp: '123',
             });
-            foo.connectFromPublic(80);
+            foo.allowFromPublic(80);
             deployment.deploy([foo]);
             expect(deploy).to.throw(
                 'foo2 has a floating IP and multiple containers. This is ' +